rename files, put back original convention for parsing functions and fix accordingly examples and test
