+++ examples/fjsp/fjsp_cpsat_example.py
@@ -4,26 +4,23 @@
 
 import logging
 
-from discrete_optimization.fjsp.flex_job_shop_parser import (
-    get_data_available_fjsp,
-    parse_fjs_file,
-)
+import discrete_optimization.fjsp.flex_job_shop_parser as fjsp_parser
+import discrete_optimization.jsp.job_shop_parser as jsp_parser
 from discrete_optimization.fjsp.flex_job_shop_problem import Job
-from discrete_optimization.fjsp.solvers.cpsat_fjsp_problem import (
+from discrete_optimization.fjsp.solvers.cpsat_fjsp_solver import (
     CPSatFJspSolver,
     FJobShopProblem,
 )
 from discrete_optimization.generic_tools.cp_tools import ParametersCP
-from discrete_optimization.jsp.job_shop_parser import get_data_available, parse_file
 from discrete_optimization.jsp.job_shop_problem import JobShopProblem, Subjob
 
 logging.basicConfig(level=logging.INFO)
 
 
 def run_cpsat_jsp():
-    file_path = get_data_available()[1]
+    file_path = jsp_parser.get_data_available()[1]
     # file_path = [f for f in get_data_available() if "abz6" in f][0]
-    problem: JobShopProblem = parse_file(file_path)
+    problem: JobShopProblem = jsp_parser.parse_file(file_path)
     fproblem = FJobShopProblem(
         list_jobs=[
             Job(job_id=i, sub_jobs=[[sj] for sj in problem.list_jobs[i]])
@@ -40,11 +37,11 @@ def run_cpsat_jsp():
 
 
 def run_cpsat_fjsp():
-    files = get_data_available_fjsp()
+    files = fjsp_parser.get_data_available()
     print(files)
     file = [f for f in files if "Behnke60.fjs" in f][0]
     print(file)
-    problem = parse_fjs_file(file)
+    problem = fjsp_parser.parse_file(file)
     print(problem)
     solver = CPSatFJspSolver(problem=problem)
     p = ParametersCP.default_cpsat()
