+++ tests/fjsp/test_fjsp_lns.py
@@ -5,11 +5,11 @@
 import logging
 
 from discrete_optimization.fjsp.flex_job_shop_parser import (
-    get_data_available_fjsp,
-    parse_fjs_file,
+    get_data_available,
+    parse_file,
 )
-from discrete_optimization.fjsp.solvers.cpsat_fjsp_problem import CPSatFJspSolver
-from discrete_optimization.fjsp.solvers.cpsat_lns_fjsp import (
+from discrete_optimization.fjsp.solvers.cpsat_fjsp_solver import CPSatFJspSolver
+from discrete_optimization.fjsp.solvers.cpsat_lns_fjsp_solver import (
     ConstraintHandlerFJSP,
     ConstraintHandlerNeighFJSP,
     NeighborBuilderSubPart,
@@ -22,10 +22,10 @@ from discrete_optimization.generic_tools.lns_tools import ConstraintHandlerMix
 
 
 def test_lnscpsat_fjsp():
-    files = get_data_available_fjsp()
+    files = get_data_available()
     file = [f for f in files if "Behnke1.fjs" in f][0]
     print(file)
-    problem = parse_fjs_file(file)
+    problem = parse_file(file)
     solver = CPSatFJspSolver(problem=problem)
     p = ParametersCP.default_cpsat()
     p.nb_process = 10
